requirements.txt
aspose-pdf==24.5.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
et-xmlfile==1.1.0
fire==0.6.0
Flask==3.0.3
fonttools==4.53.0
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.2.2
MarkupSafe==2.1.5
numpy==2.0.0
opencv-python-headless==4.10.0.84
openpyxl==3.1.4
packaging==24.1
pandas==2.2.2
pdf2docx==0.5.8
pdf2image==1.17.0
pdf2pptx==1.0.5
pillow==10.3.0
pyjnius==1.6.1
PyMuPDF==1.20.1
PyMuPDFb==1.24.3
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-pptx==0.6.23
pytz==2024.1
six==1.16.0
tabula-py==2.9.3
termcolor==2.4.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.3
XlsxWriter==3.2.0
